moved video container
